Visualize IADs debug

+++ iad_visualization.py
@@ -65,7 +65,8 @@ def run(lfd_params, model):
 
                 new_frames = []
                 for f, frame in enumerate(rgb_image):
-                    iad_frame = np.uint8(iad_img[:, f])
+                    #iad_frame = np.uint8(iad_img[:, f])
+                    iad_frame = iad_img[:, f]
                     print("2.1:", iad[0:])
                     iad_frame = iad_frame.reshape(-1, 1)
                     iad_frame = Image.fromarray(iad_frame)
